requirements.txt
python-telegram-bot==21.6
fastapi==0.115.2
uvicorn[standard]==0.31.1
